Adaptive breakpoints 0.0.3 release
