+++ adaptive_breakpoints/CHANGELOG.md
@@ -1,5 +1,9 @@
 # [0.0.2] - August 24, 2020
 
+Removed adaptive container widget.
+
+# [0.0.2] - August 24, 2020
+
 Added operator logic to AdaptiveWindowType.
 
 # [0.0.1] - August 12, 2020
