+++ adaptive_breakpoints/lib/adaptive_breakpoints.dart
@@ -5,4 +5,3 @@
 library adaptive_breakpoints;
 
 export 'src/adaptive_breakpoints.dart';
-export 'src/adaptive_container.dart';
