+++ adaptive_breakpoints/pubspec.yaml
@@ -1,6 +1,6 @@
 name: adaptive_breakpoints
 description: A Flutter package for adaptive breakpoints following the Material guidelines.
-version: 0.0.2
+version: 0.0.3
 homepage: https://github.com/material-components/material-components-flutter-adaptive
 
 environment:
